requirements/default.txt
celery==3.1.23
django-timezone-field==1.2
